Organize the order that the logstash configuration files are loaded
The configurations in logstash.conf are the most generic ones and, because
of that, must be the last to be loaded. Logstash reads the files in alphabetical
order, so, now, we prefix the filenames with numbers to "force" the order they
will be loaded. The logstash.conf file should be the only one prefixed with 9 as
a way to guarantee that it will be the last loaded file.